+++ frontend/peerprep/utils/utils.tsx
@@ -26,32 +26,10 @@ export const LANGUAGE_VERSIONS = {
 export const languages = [
   "TYPESCRIPT",
   "JAVASCRIPT",
-  "CSS",
-  "LESS",
-  "SCSS",
-  "JSON",
-  "HTML",
-  "XML",
   "PHP",
   "CSHARP",
-  "CPP",
-  "RAZOR",
-  "MARKDOWN",
-  "DIFF",
   "JAVA",
-  "VB",
-  "COFFEESCRIPT",
-  "HANDLEBARS",
-  "BATCH",
-  "PUG",
-  "FSHARP",
-  "LUA",
-  "POWERSHELL",
   "PYTHON",
-  "RUBY",
-  "SASS",
-  "R",
-  "OBJECTIVE-C",
 ];
 
 export const delay = (ms: number) =>
